requirements.txt
fastapi==0.115.6
faiss-cpu==1.9.0.post1
PyPDF2==3.0.1
python-dotenv==1.0.1
python-multipart==0.0.20
langchain==0.3.15
langchain-community==0.3.15
langchain-google-genai==2.0.9
langchain-text-splitters==0.3.5
uvicorn==0.34.0